+++ LeeSangSu-assignment/Extensions/UICollectionViewCell+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  UICollectionViewCell+Extension.swift
+//  LeeSangSu-assignment
+//
+//  Created by 이상수 on 11/12/25.
+//
+
+import UIKit
+
+extension UICollectionViewCell {
+    static var identifier : String {
+        return String(describing: self)
+    }
+}
