+++ LeeSangSu-assignment/Extensions/UITableViewCell+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  UITableViewCell+Extension.swift
+//  LeeSangSu-assignment
+//
+//  Created by 이상수 on 11/12/25.
+//
+
+import UIKit
+
+extension UITableViewCell {
+    static var identifier : String {
+        return String(describing: self)
+    }
+}
