+++ LeeSangSu-assignment/Models/CategoryItem.swift
@@ -1,0 +1,13 @@
+//
+//  CategoryItem.swift
+//  LeeSangSu-assignment
+//
+//  Created by 이상수 on 11/12/25.
+//
+
+import UIKit
+
+struct CategoryItem {
+    let image: UIImage?
+    let text: String
+}
