+++ LeeSangSu-assignment/Models/MockData.swift
@@ -1,0 +1,53 @@
+//
+//  MockData.swift
+//  LeeSangSu-assignment
+//
+//  Created by 이상수 on 11/12/25.
+//
+
+import UIKit
+
+enum MockData {
+    static var advertisementData: [UIImage?] = [
+        UIImage(systemName: "person.fill"),
+        UIImage(systemName: "person")
+    ]
+
+    static var marketData: [CategoryItem] = [
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person"), text: "이마트"),
+        CategoryItem(image: UIImage(systemName: "sun.min"), text: "홈플러스"),
+        CategoryItem(image: UIImage(systemName: "sun.max"), text: "배민"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person"), text: "이마트"),
+        CategoryItem(image: UIImage(systemName: "sun.min"), text: "홈플러스"),
+        CategoryItem(image: UIImage(systemName: "sun.max"), text: "배민"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU")
+    ]
+    
+    static var categoryData: [CategoryItem] = [
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person"), text: "이마트"),
+        CategoryItem(image: UIImage(systemName: "sun.min"), text: "홈플러스"),
+        CategoryItem(image: UIImage(systemName: "sun.max"), text: "배민"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person"), text: "이마트"),
+        CategoryItem(image: UIImage(systemName: "sun.min"), text: "홈플러스"),
+        CategoryItem(image: UIImage(systemName: "sun.max"), text: "배민"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
+        CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU")
+    ]
+}
