+++ LeeSangSu-assignment/ViewControllers/BaeminFeedViewController.swift
@@ -21,8 +21,8 @@ final class BaeminFeedViewController: UIViewController {
         view.addSubview(tableView)
         tableView.dataSource = self
         tableView.delegate = self
-        tableView.register(AdvertisementCell.self, forCellReuseIdentifier: "AdvertisementCell")
-        tableView.register(MarketCell.self, forCellReuseIdentifier: "MarketCell")
+        tableView.register(AdvertisementCell.self, forCellReuseIdentifier: AdvertisementCell.identifier)
+        tableView.register(CategoryCell.self, forCellReuseIdentifier: CategoryCell.identifier)
     }
     
     private func layoutTableView() {
@@ -34,7 +34,7 @@ final class BaeminFeedViewController: UIViewController {
 
 extension BaeminFeedViewController: UITableViewDataSource {
     func numberOfSections(in tableView: UITableView) -> Int {
-            2
+        2
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -42,26 +42,14 @@ extension BaeminFeedViewController: UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        if indexPath.section == 1 {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "AdvertisementCell", for: indexPath) as? AdvertisementCell else {
-                return UITableViewCell()
-            }
-            cell.configure(items: [UIImage(systemName: "person.fill"), UIImage(systemName: "person")])
+        switch indexPath.section {
+        case 0:
+            let cell = tableView.dequeueReusableCell(withIdentifier: CategoryCell.identifier, for: indexPath)
+            (cell as? CategoryCell)?.configure(items: MockData.marketData)
             return cell
-        } else {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "MarketCell", for: indexPath) as? MarketCell else {
-                return UITableViewCell()
-            }
-            cell.configure(items: [
-                CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
-                CategoryItem(image: UIImage(systemName: "person"), text: "이마트"),
-                CategoryItem(image: UIImage(systemName: "sun.min"), text: "홈플러스"),
-                CategoryItem(image: UIImage(systemName: "sun.max"), text: "배민"),
-                CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
-                CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
-                CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU"),
-                CategoryItem(image: UIImage(systemName: "person.fill"), text: "CU")
-            ])
+        default:
+            let cell = tableView.dequeueReusableCell(withIdentifier: AdvertisementCell.identifier, for: indexPath)
+            (cell as? AdvertisementCell)?.configure(items: MockData.advertisementData)
             return cell
         }
     }
@@ -69,6 +57,11 @@ extension BaeminFeedViewController: UITableViewDataSource {
 
 extension BaeminFeedViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        100
+        switch indexPath.section {
+        case 0:
+            return 100
+        default:
+            return 100
+        }
     }
 }
