+++ LeeSangSu-assignment/Views/CategoryCell.swift
@@ -8,50 +8,25 @@
 import UIKit
 import SnapKit
 
-final class CategoryCell: UICollectionViewCell {
-    private let icon = UIImageView()
-    private let title = UILabel()
+final class CategoryCell: BaseHorizontalCell<CategoryItem, CategoryItemCell> {
+    override var lineSpacing: CGFloat { 5 }
+    override var interitemSpacing: CGFloat { 5 }
+    override func configureCell(_ cell: CategoryItemCell, item: CategoryItem) {
+        cell.configure(image: item.image, text: item.text)
+    }
     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupIcon()
-        setupTitle()
-        layout()
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupLayout()
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    private func setupIcon() {
-        contentView.addSubview(icon)
-        icon.contentMode = .scaleAspectFit
-        icon.layer.cornerRadius = 5
-        icon.layer.borderWidth = 2
-        icon.backgroundColor = .baeminGray200
-    }
-    
-    private func setupTitle() {
-        contentView.addSubview(title)
-        title.font = .systemFont(ofSize: 10, weight: .semibold)
-    }
-    
-    private func layout() {
-        icon.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(5)
-            $0.centerX.equalToSuperview()
-            $0.size.equalTo(40)
-        }
-        title.snp.makeConstraints {
-            $0.top.equalTo(icon.snp.bottom).offset(5)
-            $0.centerX.equalToSuperview()
-        }
-    }
-}
 
-extension CategoryCell {
-    func configure(image: UIImage?, text: String) {
-        icon.image = image
-        title.text = text
+    private func setupLayout() {
+        collectionView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
     }
 }
